+++ projects/demo-maps/src/app/components/raster-feature-info/raster-feature-info.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RasterFeatureInfoComponent } from './raster-feature-info.component';
+
+describe('RasterFeatureInfoComponent', () => {
+  let component: RasterFeatureInfoComponent;
+  let fixture: ComponentFixture<RasterFeatureInfoComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ RasterFeatureInfoComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(RasterFeatureInfoComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});